Odradjen frontend deo za menadzera restorana i odgovarajuci backend.
Uvedena SystemUser klasa, radi mogucnosti cuvanja svih email-ova na
jednom mestu, radi efikasnijeg logovanja.

+++ src/main/java/com/example/model/TableRestaurant.java
@@ -10,7 +10,7 @@ import javax.persistence.ManyToOne;
 @Entity
 public class TableRestaurant {
 	@Id
-	private int tableNumber;
+	private long tableNumber;
 	
 	@Column(nullable = false)
 	private String segment;
@@ -41,11 +41,11 @@ public class TableRestaurant {
 		this.y = y;
 	}
 
-	public int getTableNumber() {
+	public long getTableNumber() {
 		return tableNumber;
 	}
 
-	public void setTableNumber(int tableNumber) {
+	public void setTableNumber(long tableNumber) {
 		this.tableNumber = tableNumber;
 	}
